+++ src/main/java/com/ysf/local/cucumberImplementataion/common/DataSeed.java
@@ -1,5 +1,6 @@
 package com.ysf.local.cucumberImplementataion.common;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import javax.annotation.PostConstruct;
@@ -22,7 +23,7 @@ public class DataSeed {
 	public void createsAccounts() {
 		for (int i = 0; i < 20; i++) {
 			repos.save(new Account(faker.number().randomNumber() , faker.name().name() 
-									,faker.random().nextDouble()+1000 ,LocalDateTime.now() ));
+									,BigDecimal.valueOf(faker.random().nextDouble()+1000) ,LocalDateTime.now() ));
 		}
 	}
 }
